+++ src/main/java/pe/com/tatapizzeria/repository/CategoriasProductoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import pe.com.tatapizzeria.entity.CategoriasProductoEntity;
 import java.util.List;
 
-public interface CategoriasProductoRepository extends JpaRepository<CategoriasProductoEntity, Long> {
+public interface CategoriasProductoRepository extends JpaRepository<CategoriasProductoEntity, Integer> {
 
     @Query("SELECT s FROM CategoriasProductoEntity s WHERE s.estado = true")
     List<CategoriasProductoEntity> findAllCustom();
