+++ src/main/java/pe/com/tatapizzeria/repository/CombosPromocionesRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import pe.com.tatapizzeria.entity.CombosPromocionesEntity;
 import java.util.List;
 
-public interface CombosPromocionesRepository extends JpaRepository<CombosPromocionesEntity, Long> {
+public interface CombosPromocionesRepository extends JpaRepository<CombosPromocionesEntity, Integer> {
 
     @Query("SELECT s FROM CombosPromocionesEntity s WHERE s.estado = true")
     List<CombosPromocionesEntity> findAllCustom();
